fix: 角色服务层调整

+++ src/modules/system/service/impl/SysRoleServiceImpl.ts
@@ -1,5 +1,4 @@
 import { Provide, Inject, Singleton } from '@midwayjs/decorator';
-import { ADMIN_ROLE_ID } from '../../../../framework/constants/AdminConstants';
 import { SysRole } from '../../model/SysRole';
 import { SysRoleDept } from '../../model/SysRoleDept';
 import { SysRoleMenu } from '../../model/SysRoleMenu';
@@ -69,7 +68,12 @@ export class SysRoleServiceImpl implements ISysRoleService {
   }
 
   async selectRoleById(roleId: string): Promise<SysRole> {
-    return await this.sysRoleRepository.selectRoleById(roleId);
+    if (!roleId) return null;
+    const roles = await this.sysRoleRepository.selectRoleByIds([roleId]);
+    if (roles.length > 0) {
+      return roles[0];
+    }
+    return null;
   }
 
   async checkUniqueRoleName(
@@ -98,16 +102,10 @@ export class SysRoleServiceImpl implements ISysRoleService {
     return !uniqueId;
   }
 
-  async countUserRoleByRoleId(roleId: string): Promise<number> {
-    return await this.sysUserRoleRepository.countUserRoleByRoleId(roleId);
-  }
-
   async insertRole(sysRole: SysRole): Promise<string> {
     const insertId = await this.sysRoleRepository.insertRole(sysRole);
     if (insertId && sysRole.menuIds) {
-      sysRole.roleId = insertId;
-      await this.insertRoleMenu(sysRole);
-      return insertId;
+      await this.insertRoleMenu(insertId, sysRole.menuIds);
     }
     return insertId;
   }
@@ -117,9 +115,8 @@ export class SysRoleServiceImpl implements ISysRoleService {
     const rows = await this.sysRoleRepository.updateRole(sysRole);
     if (rows > 0 && sysRole.menuIds) {
       // 删除角色与菜单关联
-      await this.sysRoleMenuRepository.deleteRoleMenuByRoleId(sysRole.roleId);
-      await this.insertRoleMenu(sysRole);
-      return rows;
+      await this.sysRoleMenuRepository.deleteRoleMenu([sysRole.roleId]);
+      await this.insertRoleMenu(sysRole.roleId, sysRole.menuIds);
     }
     return rows;
   }
@@ -129,11 +126,14 @@ export class SysRoleServiceImpl implements ISysRoleService {
    *
    * @param sysRole 角色对象
    */
-  private async insertRoleMenu(sysRole: SysRole): Promise<number> {
-    if (sysRole.menuIds && sysRole.menuIds.length <= 0) return 0;
-    const sysRoleMenus: SysRoleMenu[] = sysRole.menuIds.map(menuId => {
+  private async insertRoleMenu(
+    roleId: string,
+    menuIds: string[]
+  ): Promise<number> {
+    if (menuIds && menuIds.length <= 0) return 0;
+    const sysRoleMenus: SysRoleMenu[] = menuIds.map(menuId => {
       if (menuId) {
-        return new SysRoleMenu(sysRole.roleId, menuId);
+        return new SysRoleMenu(roleId, menuId);
       }
     });
     if (sysRoleMenus.length <= 0) return 0;
@@ -144,47 +144,49 @@ export class SysRoleServiceImpl implements ISysRoleService {
     const roleId = sysRole.roleId;
     // 删除角色与部门关联
     await this.sysRoleDeptRepository.deleteRoleDept([roleId]);
-    // 新增角色和部门信息（数据权限）
+    // 新增角色和部门信息
     if (sysRole.deptIds && sysRole.deptIds.length > 0) {
       const sysRoleDepts: SysRoleDept[] = sysRole.deptIds.map(deptId => {
         if (deptId) {
           return new SysRoleDept(roleId, deptId);
         }
       });
-      if (sysRoleDepts.length > 0) {
-        await this.sysRoleDeptRepository.batchRoleDept(sysRoleDepts);
-      }
+      await this.sysRoleDeptRepository.batchRoleDept(sysRoleDepts);
     }
     // 修改角色信息
     return await this.sysRoleRepository.updateRole(sysRole);
   }
 
   async deleteRoleByIds(roleIds: string[]): Promise<number> {
-    for (const roleId of roleIds) {
-      // 检查是否管理员角色
-      if (roleId === ADMIN_ROLE_ID) {
-        throw new Error('不允许操作管理员角色');
+    const roles = await this.sysRoleRepository.selectRoleByIds(roleIds);
+    if (roles.length <= 0) {
+      throw new Error('没有权限访问角色数据！');
+    }
+    for (const role of roles) {
+      // 检查是否为已删除
+      if (role.delFlag === '1') {
+        throw new Error(`${role.roleName} 角色信息已经删除！`);
       }
-      // 检查是否存在
-      const role = await this.selectRoleById(roleId);
-      if (!role) {
-        throw new Error('没有权限访问角色数据！');
-      }
-      const useCount = await this.countUserRoleByRoleId(roleId);
+      // 检查分配用户
+      const useCount = await this.sysUserRoleRepository.countUserRoleByRoleId(
+        role.roleId
+      );
       if (useCount > 0) {
         throw new Error(`【${role.roleName}】已分配给用户,不能删除`);
       }
     }
-
-    // 删除角色与菜单关联
-    await this.sysRoleMenuRepository.deleteRoleMenu(roleIds);
-    // 删除角色与部门关联
-    await this.sysRoleDeptRepository.deleteRoleDept(roleIds);
-    return await this.sysRoleRepository.deleteRoleByIds(roleIds);
+    if (roles.length === roleIds.length) {
+      // 删除角色与菜单关联
+      await this.sysRoleMenuRepository.deleteRoleMenu(roleIds);
+      // 删除角色与部门关联
+      await this.sysRoleDeptRepository.deleteRoleDept(roleIds);
+      return await this.sysRoleRepository.deleteRoleByIds(roleIds);
+    }
+    return 0;
   }
 
   async deleteAuthUsers(roleId: string, userIds: string[]): Promise<number> {
-    return await this.sysUserRoleRepository.deleteUserRoleInfos(
+    return await this.sysUserRoleRepository.deleteUserRoleByRoleId(
       roleId,
       userIds
     );
